A2/backend/requirements.txt
Flask==2.0.2
Flask-JWT-Extended==4.3.0
psycopg2-binary==2.9.1
